Fix ORB-SLAM in RGB-D mode to coerce the depth type to 32 bit. Some sources provide depth as float64, and we need to reduce it.

+++ arvet_slam/systems/slam/orbslam2.py
@@ -283,7 +283,7 @@ class OrbSlam2(VisionSystem):
                 self._input_queue.put((image_utils.convert_to_grey(image.left_pixels),
                                        image_utils.convert_to_grey(image.right_pixels), timestamp))
             elif self.mode == SensorMode.RGBD:
-                self._input_queue.put((image_utils.convert_to_grey(image.pixels), image.depth, timestamp))
+                self._input_queue.put((image_utils.convert_to_grey(image.pixels), image.depth.astype(np.float32), timestamp))
 
     def finish_trial(self) -> SLAMTrialResult:
         """
